add frontend service

+++ internal/mify/service/generate/openapi_client.go
@@ -8,6 +8,7 @@ import (
 	"unicode"
 
 	"github.com/chebykinn/mify/internal/mify/core"
+	"github.com/chebykinn/mify/internal/mify/service/lang"
 )
 
 func (g *OpenAPIGenerator) makeClientEnrichedSchema(ctx *core.Context, schemaPath string) (string, error) {
@@ -48,17 +49,21 @@ func (g *OpenAPIGenerator) doGenerateClient(ctx *core.Context, assetsPath string
 		return fmt.Errorf("failed to get client port: %w", err)
 	}
 
-	err = runOpenapiGenerator(ctx, g.basePath, schemaPath, assetsPath, generatedPath, packageName, clientPort, g.info)
+	err = runOpenapiGenerator(ctx, g.basePath, schemaPath, assetsPath,
+		generatedPath, packageName, clientName, clientPort, g.info)
 	if err != nil {
 		return fmt.Errorf("failed to run openapi-generator: %w", err)
 	}
 
-	err = os.Remove(filepath.Join(generatedPath, "api"))
-	if err != nil {
-		return err
+	// TODO: go specific
+	if g.language == lang.ServiceLanguageGo {
+		err = os.Remove(filepath.Join(generatedPath, "api"))
+		if err != nil {
+			return err
+		}
 	}
 
-	err = formatGenerated(generatedPath)
+	err = formatGenerated(generatedPath, g.language)
 	if err != nil {
 		return err
 	}
